Refactor pings structure, make buildDimensions generic
