+++ src/main/scala/com/mozilla/telemetry/streaming/ErrorAggregator.scala
@@ -203,51 +203,23 @@ object ErrorAggregator {
       .drop("window")
   }
 
-  private def buildDimensions(dimensionsSchema: StructType, meta: Meta, application: Application): Array[Row] = {
+  private def buildDimensions(dimensionsSchema: StructType, ping: Ping): Array[Row] = {
+    val meta = ping.meta
 
-    // add a null experiment_id and experiment_branch for each ping
-    val experiments = (meta.experiments :+ (None, None)).toSet.toArray
-
-    experiments.map{ case (experiment_id, experiment_branch) =>
-      val dimensions = new RowBuilder(dimensionsSchema)
-      dimensions("timestamp") = Some(meta.normalizedTimestamp())
-      dimensions("submission_date_s3") = Some(LocalDateTime.fromDateFields(meta.normalizedTimestamp()).toString(dateFormat))
-      dimensions("channel") = Some(meta.normalizedChannel)
-      dimensions("version") = meta.`environment.build`.flatMap(_.version)
-      dimensions("display_version") = application.displayVersion
-      dimensions("build_id") = meta.normalizedBuildId
-      dimensions("application") = Some(meta.appName)
-      dimensions("os_name") = meta.`environment.system`.map(_.os.name)
-      dimensions("os_version") = meta.`environment.system`.map(_.os.normalizedVersion)
-      dimensions("architecture") = meta.`environment.build`.flatMap(_.architecture)
-      dimensions("country") = Some(meta.geoCountry)
-      dimensions("experiment_id") = experiment_id
-      dimensions("experiment_branch") = experiment_branch
-      dimensions.build
-    }
-  }
-
-  private def buildDimensions(dimensionsSchema: StructType, corePing: CorePing): Array[Row] = {
-    val meta = corePing.meta
-    // add a null experiment_id and experiment_branch for each ping
-    // add a null experiment_branch for each experiment, as experiments on Fennec do not report branches
-    val experiments = (
-      corePing.experiments.map(_.map(e => (Option(e), Option.empty[String])))
-        .getOrElse(Array[(Option[String], Option[String])]()) :+ (None, None)
-      ).distinct
+    val experiments = ping.getExperiments
 
     experiments.map { case (experiment_id, experiment_branch) =>
       val dimensions = new RowBuilder(dimensionsSchema)
       dimensions("timestamp") = Some(meta.normalizedTimestamp())
       dimensions("submission_date_s3") = Some(LocalDateTime.fromDateFields(meta.normalizedTimestamp()).toString(dateFormat))
       dimensions("channel") = Some(meta.normalizedChannel)
-      dimensions("version") = Option(corePing.meta.appVersion)
-      dimensions("display_version") = Option(corePing.meta.appVersion)
+      dimensions("version") = ping.getVersion
+      dimensions("display_version") = ping.getDisplayVersion
       dimensions("build_id") = meta.normalizedBuildId
       dimensions("application") = Some(meta.appName)
-      dimensions("os_name") = Option(corePing.os)
-      dimensions("os_version") = Option(corePing.osversion)
-      dimensions("architecture") = Option(corePing.arch)
+      dimensions("os_name") = ping.getOsName
+      dimensions("os_version") = ping.getOsVersion
+      dimensions("architecture") = ping.getArchitecture
       dimensions("country") = Some(meta.geoCountry)
       dimensions("experiment_id") = experiment_id
       dimensions("experiment_branch") = experiment_branch
@@ -258,7 +230,7 @@ object ErrorAggregator {
   def parse(ping: CrashPing, dimensionsSchema: StructType, statsSchema: StructType): Array[Row] = {
     if (ping.isMainCrash || ping.isContentCrash) {
 
-      val dimensions = buildDimensions(dimensionsSchema, ping.meta, ping.application)
+      val dimensions = buildDimensions(dimensionsSchema, ping)
       val stats = new RowBuilder(SchemaBuilder.merge(statsSchema, tempSchema))
 
       stats("count") = Some(1)
@@ -288,7 +260,7 @@ object ErrorAggregator {
     val usageHours = ping.usageHours
     if (usageHours.isEmpty) throw new Exception("Main pings should have a  number of usage hours != 0")
 
-    val dimensions = buildDimensions(dimensionsSchema, ping.meta, ping.application)
+    val dimensions = buildDimensions(dimensionsSchema, ping)
     val stats = new RowBuilder(SchemaBuilder.merge(statsSchema, tempSchema))
     stats("count") = Some(1)
     stats("subsession_count") = Some(1)
